+++ python/mirascope/llm/clients/openai/completions/__init__.py
@@ -1,11 +1,9 @@
-from . import _utils
 from .clients import OpenAICompletionsClient, client, get_client
 from .model_ids import OpenAICompletionsModelId
 
 __all__ = [
     "OpenAICompletionsClient",
     "OpenAICompletionsModelId",
-    "_utils",
     "client",
     "get_client",
 ]
