+++ python/mirascope/llm/clients/openai/completions/_utils/encode.py
@@ -1,56 +1,28 @@
-"""OpenAI message types and conversion utilities."""
+"""OpenAI completions message encoding and request preparation."""
 
-import logging
 from collections.abc import Sequence
 from functools import lru_cache
-from typing import Literal
+from typing import TypedDict
 
-from openai import AsyncStream, NotGiven, Stream
+from openai import NotGiven
 from openai.types import chat as openai_types, shared_params as shared_openai_types
 from openai.types.shared_params.response_format_json_schema import JSONSchema
 
-from ....content import (
-    AssistantContentPart,
-    Text,
-    TextChunk,
-    TextEndChunk,
-    TextStartChunk,
-    ToolCall,
-    ToolCallChunk,
-    ToolCallEndChunk,
-    ToolCallStartChunk,
-)
-from ....exceptions import FormattingModeNotSupportedError
-from ....formatting import (
+from .....exceptions import FormattingModeNotSupportedError
+from .....formatting import (
     Format,
     FormattableT,
     _utils as _formatting_utils,
     resolve_format,
 )
-from ....messages import AssistantMessage, Message, UserMessage
-from ....responses import (
-    AsyncChunkIterator,
-    ChunkIterator,
-    FinishReason,
-    FinishReasonChunk,
-    RawStreamEventChunk,
-)
-from ....tools import FORMAT_TOOL_NAME, BaseToolkit, ToolSchema
-from ...base import BaseKwargs, Params, _utils as _base_utils
-from ..shared import (
-    _utils as _shared_utils,
-)
-from .model_ids import OpenAICompletionsModelId
-
-logger = logging.getLogger(__name__)
-
-OPENAI_FINISH_REASON_MAP = {
-    "length": FinishReason.MAX_TOKENS,
-    "content_filter": FinishReason.REFUSAL,
-}
+from .....messages import AssistantMessage, Message, UserMessage
+from .....tools import FORMAT_TOOL_NAME, BaseToolkit, ToolSchema
+from ....base import Params, _utils as _base_utils
+from ...shared import _utils as _shared_utils
+from ..model_ids import OpenAICompletionsModelId
 
 
-class ChatCompletionCreateKwargs(BaseKwargs, total=False):
+class ChatCompletionCreateKwargs(TypedDict, total=False):
     """Kwargs for OpenAI ChatCompletion.create method."""
 
     model: str
@@ -218,7 +190,7 @@ def _create_strict_response_format(
     )
 
 
-def prepare_completions_request(
+def encode_request(
     *,
     model_id: OpenAICompletionsModelId,
     messages: Sequence[Message],
@@ -227,9 +199,11 @@ def prepare_completions_request(
     params: Params,
 ) -> tuple[Sequence[Message], Format[FormattableT] | None, ChatCompletionCreateKwargs]:
     """Prepares a request for the `OpenAI.chat.completions.create` method."""
-    kwargs: ChatCompletionCreateKwargs = {
-        "model": model_id,
-    }
+    kwargs: ChatCompletionCreateKwargs = ChatCompletionCreateKwargs(
+        {
+            "model": model_id,
+        }
+    )
     encode_thoughts = False
 
     with _base_utils.ensure_all_params_accessed(
@@ -300,158 +274,3 @@ def prepare_completions_request(
     kwargs["messages"] = encoded_messages
 
     return messages, format, kwargs
-
-
-def decode_response(
-    response: openai_types.ChatCompletion,
-    model_id: OpenAICompletionsModelId,
-) -> tuple[AssistantMessage, FinishReason | None]:
-    """Convert OpenAI ChatCompletion to mirascope AssistantMessage."""
-    choice = response.choices[0]
-    message = choice.message
-    refused = False
-
-    parts: list[AssistantContentPart] = []
-    if message.content:
-        parts.append(Text(text=message.content))
-    if message.refusal:
-        parts.append(Text(text=message.refusal))
-        refused = True
-    if message.tool_calls:
-        for tool_call in message.tool_calls:
-            if tool_call.type == "custom":
-                # This should never happen, because we never create "custom" tools
-                # https://platform.openai.com/docs/guides/function-calling#custom-tools
-                raise NotImplementedError("OpenAI custom tools are not supported.")
-            parts.append(
-                ToolCall(
-                    id=tool_call.id,
-                    name=tool_call.function.name,
-                    args=tool_call.function.arguments,
-                )
-            )
-
-    finish_reason = (
-        FinishReason.REFUSAL
-        if refused
-        else OPENAI_FINISH_REASON_MAP.get(choice.finish_reason)
-    )
-
-    assistant_message = AssistantMessage(
-        content=parts,
-        provider="openai:completions",
-        model_id=model_id,
-        raw_content=[],
-    )
-
-    return assistant_message, finish_reason
-
-
-class _OpenAIChunkProcessor:
-    """Processes OpenAI chat completion chunks and maintains state across chunks."""
-
-    def __init__(self) -> None:
-        self.current_content_type: Literal["text", "tool_call"] | None = None
-        self.current_tool_index: int | None = None
-        self.refusal_encountered = False
-
-    def process_chunk(self, chunk: openai_types.ChatCompletionChunk) -> ChunkIterator:
-        """Process a single OpenAI chunk and yield the appropriate content chunks."""
-        yield RawStreamEventChunk(raw_stream_event=chunk)
-
-        choice = chunk.choices[0] if chunk.choices else None
-        if not choice:
-            return  # pragma: no cover
-
-        delta = choice.delta
-
-        content = delta.content or delta.refusal
-        if delta.refusal:
-            self.refusal_encountered = True
-        if content is not None:
-            if self.current_content_type is None:
-                yield TextStartChunk()
-                self.current_content_type = "text"
-            yield TextChunk(delta=content)
-
-        if delta.tool_calls:
-            if self.current_content_type == "text":
-                # In testing, I can't get OpenAI to emit text and tool calls in the same chunk
-                # But we handle this defensively.
-                yield TextEndChunk()  # pragma: no cover
-            self.current_content_type = "tool_call"
-
-            for tool_call_delta in delta.tool_calls:
-                index = tool_call_delta.index
-
-                if (
-                    self.current_tool_index is not None
-                    and self.current_tool_index > index
-                ):
-                    raise RuntimeError(
-                        f"Received tool data for already-finished tool at index {index}"
-                    )  # pragma: no cover
-
-                if (
-                    self.current_tool_index is not None
-                    and self.current_tool_index < index
-                ):
-                    yield ToolCallEndChunk()
-                    self.current_tool_index = None
-
-                if self.current_tool_index is None:
-                    if not tool_call_delta.function or not (
-                        name := tool_call_delta.function.name
-                    ):
-                        raise RuntimeError(
-                            f"Missing name for tool call at index {index}"
-                        )  # pragma: no cover
-
-                    self.current_tool_index = index
-                    if not (tool_id := tool_call_delta.id):
-                        raise RuntimeError(
-                            f"Missing id for tool call at index {index}"
-                        )  # pragma: no cover
-
-                    yield ToolCallStartChunk(
-                        id=tool_id,
-                        name=name,
-                    )
-
-                if tool_call_delta.function and tool_call_delta.function.arguments:
-                    yield ToolCallChunk(delta=tool_call_delta.function.arguments)
-
-        if choice.finish_reason:
-            if self.current_content_type == "text":
-                yield TextEndChunk()
-            elif self.current_content_type == "tool_call":
-                yield ToolCallEndChunk()
-            elif self.current_content_type is not None:  # pragma: no cover
-                raise NotImplementedError()
-
-            finish_reason = (
-                FinishReason.REFUSAL
-                if self.refusal_encountered
-                else OPENAI_FINISH_REASON_MAP.get(choice.finish_reason)
-            )
-            if finish_reason is not None:
-                yield FinishReasonChunk(finish_reason=finish_reason)
-
-
-def convert_openai_stream_to_chunk_iterator(
-    openai_stream: Stream[openai_types.ChatCompletionChunk],
-) -> ChunkIterator:
-    """Returns a ChunkIterator converted from an OpenAI Stream[ChatCompletionChunk]"""
-    processor = _OpenAIChunkProcessor()
-    for chunk in openai_stream:
-        yield from processor.process_chunk(chunk)
-
-
-async def convert_openai_stream_to_async_chunk_iterator(
-    openai_stream: AsyncStream[openai_types.ChatCompletionChunk],
-) -> AsyncChunkIterator:
-    """Returns an AsyncChunkIterator converted from an OpenAI AsyncStream[ChatCompletionChunk]"""
-    processor = _OpenAIChunkProcessor()
-    async for chunk in openai_stream:
-        for item in processor.process_chunk(chunk):
-            yield item
