+++ python/mirascope/llm/clients/openai/responses/__init__.py
@@ -1,11 +1,9 @@
-from . import _utils
 from .clients import OpenAIResponsesClient, client, get_client
 from .model_ids import OpenAIResponsesModelId
 
 __all__ = [
     "OpenAIResponsesClient",
     "OpenAIResponsesModelId",
-    "_utils",
     "client",
     "get_client",
 ]
