+++ python/mirascope/llm/clients/openai/responses/clients.py
@@ -141,7 +141,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             An `llm.Response` object containing the LLM-generated content.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -227,7 +227,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             An `llm.AsyncResponse` object containing the LLM-generated content.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -313,7 +313,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             A `llm.StreamResponse` object containing the LLM-generated content stream.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -326,7 +326,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
             stream=True,
         )
 
-        chunk_iterator = _utils.convert_openai_responses_stream_to_chunk_iterator(
+        chunk_iterator = _utils.decode_stream(
             openai_stream,
         )
 
@@ -400,7 +400,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             A `llm.AsyncStreamResponse` object containing the LLM-generated content stream.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -413,7 +413,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
             stream=True,
         )
 
-        chunk_iterator = _utils.convert_openai_responses_stream_to_async_chunk_iterator(
+        chunk_iterator = _utils.decode_async_stream(
             openai_stream,
         )
 
@@ -500,7 +500,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             A `llm.ContextResponse` object containing the LLM-generated content and context.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -599,7 +599,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             A `llm.AsyncContextResponse` object containing the LLM-generated content and context.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -698,7 +698,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             A `llm.ContextStreamResponse` object containing the LLM-generated content stream and context.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -711,7 +711,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
             stream=True,
         )
 
-        chunk_iterator = _utils.convert_openai_responses_stream_to_chunk_iterator(
+        chunk_iterator = _utils.decode_stream(
             openai_stream,
         )
 
@@ -804,7 +804,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
         Returns:
             A `llm.AsyncContextStreamResponse` object containing the LLM-generated content stream and context.
         """
-        messages, format, kwargs = _utils.prepare_responses_request(
+        messages, format, kwargs = _utils.encode_request(
             model_id=model_id,
             messages=messages,
             tools=tools,
@@ -817,7 +817,7 @@ class OpenAIResponsesClient(BaseClient[OpenAIResponsesModelId, OpenAI]):
             stream=True,
         )
 
-        chunk_iterator = _utils.convert_openai_responses_stream_to_async_chunk_iterator(
+        chunk_iterator = _utils.decode_async_stream(
             openai_stream,
         )
